+++ scripts/fpp_uninstall.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-echo "Removing plugin config & transactions...."
+#echo "Removing plugin config & transactions...."
 
-rm -f /home/fpp/media/config/plugin.fpp-zettle.json
-rm -f /home/fpp/media/config/plugin.fpp-zettle-transactions.json
+#rm -f /home/fpp/media/config/plugin.fpp-zettle.json
+#rm -f /home/fpp/media/config/plugin.fpp-zettle-transactions.json